Refactor share class procedures and add new share class functionality
