+++ src/trpc/routers/share-class/procedures/get-share-class.ts
@@ -1,0 +1,20 @@
+import { withAuth } from "@/trpc/api/trpc";
+
+export const getShareClassProcedure = withAuth.query(async ({ ctx }) => {
+  const {
+    db,
+    session: { user },
+  } = ctx;
+
+  const data = await db.shareClass.findMany({
+    where: {
+      companyId: user.companyId,
+    },
+
+    orderBy: {
+      createdAt: "desc",
+    },
+  });
+
+  return { data };
+});
